requirements.txt
pillow
nibabel
imageio
requests
tqdm
numpy
scipy